Add second target to hello_world module
