+++ test/test-addon.js
@@ -7,40 +7,26 @@ var addonPath = path.resolve(__dirname, 'node_modules', 'hello_world')
 var nodeGyp = path.resolve(__dirname, '..', 'bin', 'node-gyp.js')
 
 test('build simple addon', function (t) {
-  t.plan(3)
+  t.plan(4)
 
-  exec(t, ['rebuild'], [], function (err, stdout, stderr) {
+  exec(t, ['clean', 'rebuild'], [], function (err, stdout, stderr) {
     var logLines = stderr.toString().trim().split(/\r?\n/)
     var lastLine = logLines[logLines.length-1]
     t.strictEqual(err, null)
     t.strictEqual(lastLine, 'gyp info ok', 'should end in ok')
     try {
-      var binding = require('hello_world')
-      t.strictEqual(binding.hello(), 'world')
+      var hello = require('hello_world/hello')
+      t.strictEqual(hello.hello(), 'world')
     } catch (error) {
       t.error(error, 'load module')
     }
-  })
-})
 
-test('build specific addon', function (t) {
-  t.plan(4)
-
-  exec(t, ['clean', 'configure'], [], function (err) {
-    t.error(err, 'clean build')
-
-    exec(t, ['build'], ['hello'], function (err, stdout, stderr) {
-      var logLines = stderr.toString().trim().split(/\r?\n/)
-      var lastLine = logLines[logLines.length-1]
-      t.strictEqual(err, null)
-      t.strictEqual(lastLine, 'gyp info ok', 'should end in ok')
-      try {
-        var binding = require('hello_world')
-        t.strictEqual(binding.hello(), 'world')
-      } catch (error) {
-        t.error(error, 'load module')
-      }
-    })
+    try {
+      var world = require('hello_world/world')
+      t.strictEqual(world.world(), 'hello')
+    } catch (error) {
+      t.error(error, 'load module')
+    }
   })
 })
